requirements.txt
psycopg2-binary>=2.9.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
flask==2.3.3
dashscope>=1.13.6 